Invitation and registartion done

+++ src/main/java/com/example/growtogether/dto/branch/request/AddNewBranchRequestDto.java
@@ -26,7 +26,7 @@ public class AddNewBranchRequestDto {
     private Long number;
 
     @NotNull
-    private Long numberOfTables;
+    private Integer numberOfTables;
 
 
 }
